requirements.txt
amqp==2.6.1
asgiref==3.4.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==4.4.7
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.0
django-cors-headers==3.10.1
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kombu==4.6.11
MarkupSafe==2.0.1
packaging==21.3
prompt-toolkit==3.0.24
psycopg2==2.9.3
PyJWT==2.3.0
pyparsing==3.0.6
pytz==2021.3
redis==4.1.0
requests==2.26.0
ruamel.yaml==0.17.19
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
vine==1.3.0
wcwidth==0.2.5
wrapt==1.13.3
